docs: code clean-up and go routine doc
