+++ docs/goroutine.md
@@ -1,0 +1,5 @@
+# Mongo Replication - Go Routines
+
+Below a simple schema representing the current threading model in Mongo Repl
+
+![](./mongo-repl-goroutine.png)
